Fix Errors in Pity Adjustments
In adding the new 4* Special Rate pool, I made an error in getting
the pity rate adjustments, mistakenly dividing the proportion by
a further 100 which messed it up and made the 4* Special Rate seem
to have an outsize impact on the sims despite that shouldn't being
the case. This fixes this error.